requirements.txt
selenium==4.23.1
python-dotenv
openai==1.35.14
asyncio
pydub==0.25.1
pyannote.audio==3.1.1
undetected-chromedriver==3.5.5
whisper==1.1.10